Init readme , change time function
