+++ src/index.ts
@@ -1,15 +1,11 @@
+/* eslint-disable @typescript-eslint/explicit-function-return-type */
 
-import { createHandler } from './create-handler'
+import { OpenApiGenerator } from './OpenApiGenerator'
 import { SwaggerConfig } from './interfaces'
-import { runSwagger } from './functions'
+
 export { SwaggerDoc, SwaggerConfig, ResponseType } from './interfaces'
 export { RequestHandlerWithDocumentation } from './create-handler'
-export class OpenApiGenerator {
-  static swaggerConfig: SwaggerConfig
-  constructor (swaggerConfig: SwaggerConfig) {
-    OpenApiGenerator.swaggerConfig = swaggerConfig
-  }
-
-  createHandler = createHandler
-  runSwagger = runSwagger
+export function setupOpenApi (swaggerConfig: SwaggerConfig) {
+  const { createHandler, runSwagger } = new OpenApiGenerator(swaggerConfig)
+  return { createHandler, runSwagger }
 }
